services/crawler/requirements.txt
google-cloud-pubsub
google-cloud-firestore
feedparser
